Add files via upload
This is my website for my portfolio for the course "Social Media". I tried making a aesthetic pleasing website about me and my interests. Unfortunately  I could not place my personal art works in there, so I had to place some placeholders instead (smileys). The rest of the website is a summary about all the topics we did in class.